ALGO-11 Added method contains to class Stack
