+++ task-09-kosaraju-algorithm/pom.xml
@@ -16,4 +16,12 @@
         <maven.compiler.target>11</maven.compiler.target>
     </properties>
 
+    <dependencies>
+        <dependency>
+            <groupId>ru.otus</groupId>
+            <artifactId>task-04-dynamic-arrays</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+
 </project>
